requirements.txt
ax_platform==0.1.20
autots @ git+git://github.com/jambo6/autots@v0.0.5
dnspython==2.1.0
sacredex @ git+git://github.com/jambo6/sacredex@v0.0.3
matplotlib==3.3.4
#mdbh @ git+https://gitlab.com/MaxSchambach/mdbh
numpy==1.20.1
pandas==1.2.2
pbr==5.5.1
pre-commit==2.10.1
pytest==6.2.2
pytorch_ignite==0.4.4
scikit-learn==0.24.1
sktime==0.6.1
sparselinear==0.0.5
torch==1.7.1
# torch-sparse==0.6.8
# torchdiffeq==0.2.1
tqdm==4.57.0
